+++ Code/Mantid/Framework/DataObjects/inc/MantidDataObjects/PeakColumn.h
@@ -29,8 +29,8 @@ namespace DataObjects
     
 
     /// Number of individual elements in the column.
-    virtual int size() const
-    { return static_cast<int>(peaks.size()); }
+    virtual size_t size() const
+    { return peaks.size(); }
 
     /// Returns typeid for the data in the column
     virtual const std::type_info& get_type_info()const;
@@ -39,7 +39,7 @@ namespace DataObjects
     virtual const std::type_info& get_pointer_type_info()const;
 
     /// Prints
-    virtual void print(std::ostream& s, int index) const;
+    virtual void print(std::ostream& s, size_t index) const;
 
     /// Specialized type check
     virtual bool isBool()const;
@@ -49,13 +49,13 @@ namespace DataObjects
 
   protected:
     /// Sets the new column size.
-    virtual void resize(int count);
+    virtual void resize(size_t count);
     /// Inserts an item.
-    virtual void insert(int index);
+    virtual void insert(size_t index);
     /// Removes an item.
-    virtual void remove(int index);
+    virtual void remove(size_t index);
     /// Pointer to a data element
-    virtual void* void_pointer(int index);
+    virtual void* void_pointer(size_t index);
 
   private:
     /// Reference to the peaks object saved in the PeaksWorkspace.
