+++ Code/Mantid/Framework/DataObjects/src/PeaksWorkspace.cpp
@@ -101,7 +101,7 @@ namespace DataObjects
 
   //---------------------------------------------------------------------------------------------
   /// Gets the shared pointer to a column by index.
-  boost::shared_ptr<Mantid::API::Column> PeaksWorkspace::getColumn(int index)
+  boost::shared_ptr<Mantid::API::Column> PeaksWorkspace::getColumn(size_t index)
   {
     if (index < 0 || index >= static_cast<int>(columns.size())) throw std::invalid_argument("PeaksWorkspace::getColumn() called with invalid index.");
     return columns[index];
